refactor : 프로필 설정 컴포넌트 분리
