+++ components/common/Header/ProfileSetting/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfileSettingModal } from './ProfileSettingModal';
