+++ components/common/Header/index.ts
@@ -1,3 +1,2 @@
 export { default as Header } from './Header';
 export { default as SettingModal } from './SettingModal';
-export { default as ProfileSettingModal } from './ProfileSettingModal';
